requirements.txt
boto3_type_annotations
toolz
google-api-python-client
tweepy
beautifulsoup4
lxml
html5lib
requests